upd: more  changes requested in the code review
