+++ divi/src/CoinstakeCreator.cpp
@@ -8,9 +8,7 @@
 #include <SuperblockHelpers.h>
 #include <Settings.h>
 #include <BlockIncentivesPopulator.h>
-#ifndef BITCOIN_MAIN_H
 #include <main.h>
-#endif
 
 extern Settings& settings;
