+++ divi/src/init.cpp
@@ -306,7 +306,7 @@ void PrepareShutdown()
     boost::filesystem::remove(GetPidFile());
 #endif
 
-    UnregisterAllValidationInterfaces(&registry);
+    UnregisterAllValidationInterfaces();
 }
 
 /**
@@ -1616,7 +1616,7 @@ bool InitializeDivi(boost::thread_group& threadGroup)
         LogPrintf("%s", strErrors.str());
         LogPrintf(" wallet      %15dms\n", GetTimeMillis() - nStart);
 
-        RegisterValidationInterface(&registry,pwalletMain);
+        RegisterValidationInterface(pwalletMain);
 
         CBlockIndex* pindexRescan = chainActive.Tip();
         if (GetBoolArg("-rescan", false))
