+++ divi/src/txmempool.cpp
@@ -17,9 +17,7 @@
 
 
 #include "FeeAndPriorityCalculator.h"
-#ifndef BITCOIN_VALIDATIONSTATE_H
 #include <ValidationState.h>
-#endif
 
 using namespace std;
